+++ src/views/user/modules/user-operate-drawer.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { computed, ref, watch } from 'vue';
-import { jsonClone } from '@sa/utils';
 import { getEnableStatusLabel, getEnableStatusOptions, getEnableStatusTagType } from '@/constants/common';
 import {
   fetchCreateUser,
@@ -10,6 +9,7 @@ import {
   fetchUpdateUser
 } from '@/service/api';
 import { useAuthStore } from '@/store/modules/auth';
+import { jsonClone } from '@/utils/klona';
 import { $t } from '@/locales';
 import FormModalWrapper from '@/components/advanced/form-modal-wrapper.vue';
 import { useOperateForm } from '../../_shared/composables/use-operate-form';